# swagger.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Reconstrua API
  description: >-
    A API do Reconstrua é uma plataforma inovadora que conecta pessoas em situação de rua ou afetadas por tragédias com voluntários dispostos a ajudar. Através de nossa API, é possível gerenciar empresas, voluntários, beneficiados e administradores, facilitando o processo de doação, alojamento e apoio emocional. Nossa missão é transformar vidas através da solidariedade e da tecnologia, proporcionando moradia e esperança para aqueles que mais precisam.


    [Frontend Recontrua](https://reconstrua.netlify.app/)
...
paths:
...
  /api/beneficiaries:
    post:
      tags:
        - beneficiary
      operationId: postBeneficiary
      summary: Cadastra um beneficiado
      description: Cadastra um novo beneficiado.
      requestBody:
        description: Os dados necessários para cadastro de beneficiado devem ser especificados usando o formato JSON no corpo da requisição.
        $ref: "#/components/requestBodies/Beneficiary"
      responses:
        "201":
          description: Beneficiado cadastrado com sucesso.
          $ref: "#/components/responses/Success/Default"
        "409":
          description: Beneficiado já cadastrado.
          $ref: "#/components/responses/Error"
        "422":
          description: Pelo menos um dos valores enviado no corpo da requisição é inválido.
          $ref: "#/components/responses/Error422"
        "500":
          description: O servidor encontrou um erro ao tentar cadastrar um beneficiado.
          $ref: "#/components/responses/Error"
    get:
      tags:
        - beneficiary
      operationId: getAllBeneficiaries
      summary: Busca todos os beneficiados existentes
      description: Para utilizar essa rota é preciso estar autenticado, realize o login e insira o token de autenticação no authorize.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Voluntários buscados com sucesso e retornados no corpo da resposta.
          $ref: "#/components/responses/Success/GetBeneficiaries"
        "401":
          description: Você não está autenticado como administrador.
          $ref: "#/components/responses/Error"
        "404":
          description: Beneficiados não encontrados.
          $ref: "#/components/responses/Error"
        "500":
          description: O servidor encontrou um erro ao tentar buscar beneficiados.
          $ref: "#/components/responses/Error"
  /api/beneficiaries/{beneficiary_id}:
    get:
      tags:
        - beneficiary
      operationId: getBeneficiaryById
      summary: Busca um beneficiado
      description: Para utilizar essa rota é preciso  estar autenticado, realize o login e insira o token de autenticação no authorize.
      security:
        - bearerAuth: []
      parameters:
        - name: beneficiary_id
          in: path
          description: ID do beneficiado a ser buscado
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Beneficiado buscado com sucesso e retornado no corpo da resposta.
          $ref: "#/components/responses/Success/GetBeneficiary"
        "404":
          description: Beneficiado não encontrado.
          $ref: "#/components/responses/Error"
        "401":
          description: Você não está autenticado como administrador.
          $ref: "#/components/responses/Error"
        "422":
          description: ID de beneficiado fornecido é inválido.
          $ref: "#/components/responses/Error422"
        "500":
          description: O servidor encontrou um erro ao tentar buscar beneficiado.
          $ref: "#/components/responses/Error"
    put:
      tags:
        - beneficiary
      operationId: updateBeneficiary
      summary: Atualiza um beneficiado
      description: Para utilizar essa rota é preciso estar autenticado, realize o login e insira o token de autenticação no authorize.
      security:
        - bearerAuth: []
      parameters:
        - name: beneficiary_id
          in: path
          description: ID do beneficiado a ser atualizado
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        description: Os dados necessários para atualização de beneficiado devem ser especificados usando o formato JSON no corpo da requisição.
        $ref: "#/components/requestBodies/Beneficiary"
      responses:
        "200":
          description: Beneficiado atualizado com sucesso.
          $ref: "#/components/responses/Success/Default"
        "401":
          description: Você não está autenticado como administrador.
          $ref: "#/components/responses/Error"
        "404":
          description: Beneficiado não encontrado.
          $ref: "#/components/responses/Error"
        "422":
          description: Pelo menos um dos valores enviado no corpo ou parâmetro da requisição é inválido.
          $ref: "#/components/responses/Error422"
        "500":
          description: O servidor encontrou um erro ao tentar atualizar beneficiado.
          $ref: "#/components/responses/Error"
    delete:
      tags:
        - beneficiary
      operationId: deleteBeneficiary
      summary: Deleta um beneficiado
      description: Para utilizar essa rota é preciso estar autenticado, realize o login e insira o token de autenticação no authorize.
      security:
        - bearerAuth: []
      parameters:
        - name: beneficiary_id
          in: path
          description: ID do beneficiado a ser deletado
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Beneficiado deletado com sucesso.
          $ref: "#/components/responses/Success/Default"
        "401":
          description: Você não está autenticado como administrador.
          $ref: "#/components/responses/Error"
        "404":
          description: Beneficiado não encontrado.
          $ref: "#/components/responses/Error"
        "422":
          description: ID do beneficiado fornecido é inválido.
          $ref: "#/components/responses/Error422"
        "500":
          description: O servidor encontrou um erro ao tentar deletar beneficiado.
          $ref: "#/components/responses/Error"
  # COMPANY
...
components:
  schemas:
...
    Beneficiary:
      type: object
      properties:
        first_name:
          type: string
          format: string
          examples: ["Maria"]
        last_name:
          type: string
          format: string
          examples: ["Fernandes"]
        email:
          type: string
          format: string
          examples: ["mariafernandes22@gmail.com"]
        phone:
          type: string
          format: string
          examples: ["11 987654321"]
        cep:
          type: string
          format: string
          examples: ["04547-000"]
        street:
          type: string
          format: string
          examples: ["Rua dos Três Pinos"]
        neighborhood:
          type: string
          format: string
          examples: ["Vila Mariana"]
        city:
          type: string
          format: string
          examples: ["São Paulo"]
        state:
          type: string
          format: string
          examples: ["São Paulo"]
        address_number:
          type: number
          format: number
          examples: [123]
        description:
          type: string
          format: string
          examples: ["Maria é uma jovem mãe de dois filhos pequenos que perdeu tudo em um incêndio doméstico. Ela está procurando por uma casa segura para seus filhos e está disposta a ajudar na limpeza e manutenção da nova residência. muito dedicada e trabalha em dois empregos para sustentar sua família. Ela sonha em dar uma vida melhor para seus filhos e está ansiosa para começar a nova fase em sua vida."]
      required:
        - first_name
        - last_name
        - email
        - phone
        - cep
        - street
        - neighborhood
        - city
        - state
        - address_number
        - description
    # COMPANY
...
  requestBodies:
...
    Beneficiary:
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Beneficiary"
      required: true
    # COMPANY
...
  responses:
...
    Error:
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                format: string
                examples: ["<Mensagem de erro>."]
            required:
              - message
    Error422:
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  type: object
                  properties:
                    type:
                      type: string
                      format: string
                      example: string
                    value:
                      type: string
                      format: string
                      example: string
                    msg:
                      type: string
                      format: string
                      example: string
                    path:
                      type: string
                      format: string
                      example: string
                    location:
                      type: string
                      format: string
                      example: string

  # AUTENTICAÇÃO
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Insira o token de autenticação JWT retornado pelo login de administrador